lesson-9: code refactoring
